+++ NEWS.md
@@ -11,6 +11,8 @@
 * Added `Fire$data_store` field to suppleant `Fire$get_data()` and friends. It
   provides direct acces to the global data store
 * Allow naming plugins when they are attached
+* `showcase` in `Fire$ignite()`/`Fire$run()` can now take a string giving the
+  path to open the browser at
 
 # fiery 1.2.1